+++ spiel/app.py
@@ -190,6 +190,13 @@ class SpielApp(App[None]):
 
 
 def present(deck_path: Path | str, watch_path: Optional[Path | str] = None) -> None:
+    """
+    Present the deck defined in the given `deck_path`.
+
+    Args:
+        deck_path: The file to look for a deck in.
+        watch_path: When filesystem changes are detected below this path (recursively), reload the deck from the `deck_path`.
+    """
     os.environ["TEXTUAL"] = ",".join(sorted(["debug", "devtools"]))
 
     deck_path = Path(deck_path).resolve()